+++ l2fwd/main.c
@@ -3,7 +3,7 @@ int is_client = -1, client_id;
 volatile struct wm_queue *wmq;
 
 struct cuckoo_bucket *ht_index;
-uint32_t *mac_addrs;
+ULL *mac_addrs;
 
 static struct ether_addr l2fwd_ports_eth_addr[RTE_MAX_ETHPORTS]; /**< MACs */
 struct rte_mempool *l2fwd_pktmbuf_pool[RTE_MAX_LCORE];	/**< Per lcore mempools */
