+++ l2fwd/main.h
@@ -109,7 +109,7 @@ int get_socket_id_from_macaddr(int port_id);
 void check_all_ports_link_status(uint8_t port_num, int portmask);
 
 void run_server(volatile struct wm_queue *wmq);
-void run_client(int client_id, uint32_t *mac_addrs,
+void run_client(int client_id, unsigned long long *mac_addrs,
 	struct rte_mempool **l2fwd_pktmbuf_pool);
 
 void micro_sleep(double us, double cycles_to_ns_fac);
